+++ auth/jwt_handler.py
@@ -13,9 +13,9 @@ def token_response(token: str):
 secret_key = Settings().secret_key
 
 
-def sign_jwt(user_id: str, role: str) -> Dict[str, str]:
+def sign_jwt(employee_id: str, role: str) -> Dict[str, str]:
     # Set the expiry time.
-    payload = {"user_id": user_id, "role": role, "expires": time.time() + 2400}
+    payload = {"employee_id": employee_id, "role": role, "expires": time.time() + 2400}
     return token_response(jwt.encode(payload, secret_key, algorithm="HS256"))
